Now saving and retrieving departments in MongoDB.

+++ spring-boot/webflux-tdd-demo/src/main/java/com/pintailconsultingllc/webflux/demo/dtos/DepartmentDTO.java
@@ -6,12 +6,14 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigInteger;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class DepartmentDTO {
-    Integer id;
+    BigInteger id;
     String name;
 
     public DepartmentDTO(Department department) {
