+++ spring-boot/webflux-tdd-demo/src/main/java/com/pintailconsultingllc/webflux/demo/repositories/DepartmentRepository.java
@@ -3,5 +3,7 @@ package com.pintailconsultingllc.webflux.demo.repositories;
 import com.pintailconsultingllc.webflux.demo.entities.Department;
 import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
 
-public interface DepartmentRepository extends ReactiveMongoRepository<Department, Integer> {
+import java.math.BigInteger;
+
+public interface DepartmentRepository extends ReactiveMongoRepository<Department, BigInteger> {
 }
